This commit was manufactured by cvs2svn to create tag 'Release-1_1_7'.
git-svn-id: svn://scribus.net/tags/Release-1_1_7/Scribus@303 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/mdup.h
@@ -21,7 +21,7 @@ class Mdup : public QDialog
     Q_OBJECT
 
 public:
-    Mdup( QWidget* parent, float Dx, float Dy, int Ein );
+    Mdup( QWidget* parent, double Dx, double Dy, int Ein );
     ~Mdup() {};
 
     QLabel* TextLabel1_2;
